+++ src/module/Hooks.ts
@@ -1,5 +1,7 @@
 import { warn, error, debug, i18n } from "../main";
+import { ChatLink } from "./chatlink";
 import { ChatPortrait } from "./ChatPortrait";
+import { MessageRenderData } from "./MessageRenderData";
 import { MODULE_NAME } from "./settings";
 
 export let readyHooks = async () => {
@@ -10,7 +12,10 @@ export let readyHooks = async () => {
   * This line connects our method above with the chat rendering.
   * Note that this happens after the core code has already generated HTML.
   */
-  Hooks.on('renderChatMessage', ChatPortrait.onRenderChatMessage);
+  Hooks.on('renderChatMessage', (message, html, speakerInfo) => {
+    ChatPortrait.onRenderChatMessage(message, html, <MessageRenderData>speakerInfo, speakerInfo);
+    ChatLink.prepareEvent(message, html, speakerInfo);
+  }); 
   
 }