+++ src/module/preloadTemplates.ts
@@ -5,6 +5,7 @@ export const preloadTemplates = async function () {
 		// Add paths to "module/XXX/templates"
 		//`/modules/${MODULE_NAME}/templates/XXX.html`,
 		`/modules/${MODULE_NAME}/templates/chat-portrait-form.html`,
+		`/modules/${MODULE_NAME}/templates/instructions.html`,
 	];
 
 	return loadTemplates(templatePaths);